requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.9
Django==5.2.4
django-cors-headers==4.7.0
django-dotenv==1.4.2
djangorestframework==3.16.0
et_xmlfile==2.0.0
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
yarl==1.20.1
mediapipe==0.10.32
Pillow==12.1.1
